pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gpmaster"
version = "1.0.0"
description = "GPG-backed lockbox for secrets management"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "GPMaster Contributors"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "python-gnupg>=0.5.0",
    "pyotp>=2.8.0",
]

[project.scripts]
gpmaster = "gpmaster.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/gpmaster"
Repository = "https://github.com/yourusername/gpmaster"

[tool.setuptools.packages.find]
include = ["gpmaster*"]
exclude = ["tests*"]
